fix(sjwh): 1:1 redo rws+jhdwyds+swrksj+jczhkxwh+wjgl — dept filtering, full WHERE
rws: fix PK type Integer→Long to match V1 cxtj_rws table
jhdwyds: fix PK type Integer→Long, add dept filtering (100000→null),
  fix nsrmc/ghmc/zgswskfjmc/bz to use like (not eq), add djxh/ghmc to PageReqVO,
  fix zgrs/ghzgrs type Integer→String to match V1
swrksj: add XML mapper with V1 subquery (dept_id='000000', not in gh_hj,
  computed dept_id1 from dm_swjg), add dept filtering in service
  (100000/626xxx→null), add deptId1/createBy/updateBy/createTime/updateTime to DO
jczhkxwh: add dept filtering (100000→null), add dclghbj(@TableField HJFL8_DM)/
  xwyqbj(@TableField HJFL9_DM)/createBy/createTime/updateBy/updateTime to DO,
  lowercase SQL in mapper
wjgl: lowercase SQL in FileMapper.xml, add @TableField annotations to DO


git-svn-id: svn://172.16.27.209/gsgh/gsgh/lghjft-backend@1091 63aaa323-bd20-4782-bdcf-f624a29d9816

+++ yudao-module-lghjft/src/main/java/cn/iocoder/yudao/module/lghjft/controller/admin/sjwh/jhdwyds/JhdwydsController.java
@@ -39,7 +39,7 @@ public class JhdwydsController {
     @PostMapping("/create")
     @Operation(summary = "创建应代收单位")
     @PreAuthorize("@ss.hasPermission('lghjft:sjwh-jhdwyds:create')")
-    public CommonResult<Integer> createjhdwyds(@Valid @RequestBody JhdwydsSaveReqVO createReqVO) {
+    public CommonResult<Long> createjhdwyds(@Valid @RequestBody JhdwydsSaveReqVO createReqVO) {
         return success(JhdwydsService.createjhdwyds(createReqVO));
     }
 
@@ -55,7 +55,7 @@ public class JhdwydsController {
     @Operation(summary = "删除应代收单位")
     @Parameter(name = "id", description = "编号", required = true)
     @PreAuthorize("@ss.hasPermission('lghjft:sjwh-jhdwyds:delete')")
-    public CommonResult<Boolean> deletejhdwyds(@RequestParam("id") Integer id) {
+    public CommonResult<Boolean> deletejhdwyds(@RequestParam("id") Long id) {
         JhdwydsService.deletejhdwyds(id);
         return success(true);
     }
@@ -64,7 +64,7 @@ public class JhdwydsController {
     @Parameter(name = "ids", description = "编号", required = true)
     @Operation(summary = "批量删除应代收单位")
     @PreAuthorize("@ss.hasPermission('lghjft:sjwh-jhdwyds:delete')")
-    public CommonResult<Boolean> deletejhdwydsList(@RequestParam("ids") List<Integer> ids) {
+    public CommonResult<Boolean> deletejhdwydsList(@RequestParam("ids") List<Long> ids) {
         JhdwydsService.deletejhdwydsListByIds(ids);
         return success(true);
     }
@@ -73,7 +73,7 @@ public class JhdwydsController {
     @Operation(summary = "获得应代收单位")
     @Parameter(name = "id", description = "编号", required = true, example = "1024")
     @PreAuthorize("@ss.hasPermission('lghjft:sjwh-jhdwyds:query')")
-    public CommonResult<JhdwydsResVO> getjhdwyds(@RequestParam("id") Integer id) {
+    public CommonResult<JhdwydsResVO> getjhdwyds(@RequestParam("id") Long id) {
         JhdwydsDO jhdwyds = JhdwydsService.getjhdwyds(id);
         return success(BeanUtils.toBean(jhdwyds, JhdwydsResVO.class));
     }
