+++ yudao-module-lghjft/src/main/java/cn/iocoder/yudao/module/lghjft/dal/dataobject/sjwh/rws/RwsDO.java
@@ -25,7 +25,7 @@ public class RwsDO {
      * 任务id
      */
     @TableId
-    private Integer rwid;
+    private Long rwid;
     /**
      * 任务类型	
      *
